+++ webpack.config.cjs
@@ -88,12 +88,12 @@ module.exports = {
       shared: {
         react: {
           singleton: true,
-          requiredVersion: '>=19.2.0 <20.0.0',
+          requiredVersion: '^19.2.0',
           import: false
         },
         'react-dom': {
           singleton: true,
-          requiredVersion: '>=19.2.0 <20.0.0',
+          requiredVersion: '^19.2.0',
           import: false
         }
       }